Add hash_dataset
